Clear form on printing
If you print an epic and then want to print
another you'd likely just type into the fields.
If we stored the just printed epic in activeEpic,
that would mean that the user inadvertently
overwrote the previous values.
By clearing the form you can either safely
enter new data, or click on the previously
printed one to retrieve it.